Add files via upload
### Transient Concepts in Streaming Graphs
Concept Drift (CD) occurs when a change in a hidden context can induce changes in a target concept. CD is a natural phenomenon in streaming data due to the non-stationary setting.  Understanding, detection, and adaptation to CD in streaming data is (i) vital for effective and efficient analysis/analytics as reliable output depends on fresh input, (ii) challenging as CD detection intrinsically requires stateful yet efficient operations as well as effective mechanisms for CD evaluations given the pressing issue of drift label acquisition, and (iii) impactfull as a variety of practical use-case scenarios reside in streaming setting and incur CD. The research in CD is mostly focused on black-box drift adaptation and/or cases where data instances are assumed to be independent, and the target concept is defined based on labelled data instances wrt to a supervised downstream task. These assumptions do not always hold.  We define CD in the case of transient, interconnected, and sequential data instances forming a streaming graph which serves as the input to any online adaptive analytic task (in both supervised and unsupervised mode). We introduce a modular framework for understanding and unsupervised detection of transient concept drifts. Our framework is composed of independent components with explainable, and adaptive techniques for data management and drift detection based on the inter-connectivity patterns of streaming bicliques. Our evaluations of effectiveness and efficiency confirm that introduced approaches can achieve true detection accuracy of up to $96\%$ in streams with reoccurring drift pattern and $100\%$ in streams with gradual drift pattern, zero false positive and duplicate detection rate, almost zero miss detection rates in detecting reoccurring/gradual drifts, average detection latency below $100$ (s) when the stream is bursty and $1.5$ (s) when it has low burstiness.

### Code

Use ConceptDrift.java to run the framework. To do so, set the variable filename as the name of input stream data in the main.
Make sure to include the BipartiteGraph.java in the same directory as the ConceptDrift.java. This file include classes for the Bipartite Graph (BBG) and Weighted Unipartite Graph (UWGO) data models and extra helper methods.

Use sGrow_StreamWithDrift.java to regenerate the synthetic graph streams.

Use Matlab code figurecode.m to regenerate the plots/figures given the files including the data points. The data files for figures include 4 and 3 columns for Gradual and Reoccurring drift patterns: Row#, Delay1, Delay2, and Delay3.


### Data

20 synthetic graph streams are used in the paper which will be availale upon a request. Each sGrow graph stream includes 4 columns: user, item, rating (unused), and timestamp. All streams include a  prefix of 1000 sgrs from  user-item stream (available at public repositories KONECT https://konect.cc/ and Netzschleuder https://networks.skewed.de/). Detailed information on data generation scheme are available in the paper.

Jérôme Kunegis. 2013. Konect: the koblenz network collection. In Proc. 22nd Int.World Wide Web Conf. 1343–1350.

Tiago P. Peixoto. 2020. The Netzschleuder network catalogue and repository.